Settings refactor, CippCallout wip
